fix(pipeline_template): Build partials, execute conditionals after and trim conditions as the last step
