+++ orca-pipelinetemplate/src/test/resources/templates/conditional-partials.yml
@@ -1,0 +1,26 @@
+schema: "1"
+id: simpleTemplate
+variables:
+- name: includeWait
+  type: boolean
+stages:
+- id: stageWithPartialsAndConditional
+  type: partial.partialWithConditional
+  dependsOn: []
+  config: {}
+
+partials:
+- id: partialWithConditional
+  usage: Partial that conditionally adds a step
+  variables: []
+  stages:
+  - id: conditionalWaitOnPartial
+    type: wait
+    config:
+      waitTime: 5
+    when:
+      - "{{ includeWait }}"
+  - id: wait
+    type: wait
+    config:
+      waitTime: 5
